+++ apps/backend/src/app/api/latest/internal/metrics/route.tsx
@@ -1,14 +1,13 @@
-import { Tenancy } from "@/lib/tenancies";
 import { getOrSetCacheValue } from "@/lib/cache";
+import { Tenancy } from "@/lib/tenancies";
 import { getPrismaClientForTenancy, getPrismaSchemaForTenancy, globalPrismaClient, PrismaClientTransaction, sqlQuoteIdent } from "@/prisma-client";
 import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
+import { Prisma } from "@prisma/client";
 import { KnownErrors } from "@stackframe/stack-shared";
 import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
 import { adaptSchema, adminAuthTypeSchema, yupArray, yupMixed, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import yup from 'yup';
-import { userFullInclude, userPrismaToCrud } from "../../users/crud";
-import { usersCrudHandlers } from "../../users/crud";
-import { Prisma } from "@prisma/client";
+import { userFullInclude, userPrismaToCrud, usersCrudHandlers } from "../../users/crud";
 
 type DataPoints = yup.InferType<typeof DataPointsSchema>;
