+++ apps/dashboard/src/lib/prefetch/url-prefetcher.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useAdminApp } from "@/app/(main)/(protected)/projects/[projectId]/use-admin-app";
-import { useUser } from "@stackframe/stack";
+import { stackAppInternalsSymbol, useUser } from "@stackframe/stack";
 import { previewTemplateSource } from "@stackframe/stack-shared/dist/helpers/emails";
 import { createCachedRegex } from "@stackframe/stack-shared/dist/utils/regex";
 import { useEffect, useState } from "react";
@@ -11,8 +11,12 @@ import { HookPrefetcher, HookPrefetcherCallback } from "./hook-prefetcher";
 // this is because we suspend the component
 const urlPrefetchers: Record<string, ((match: RegExpMatchArray, query: URLSearchParams, hash: string) => void | HookPrefetcherCallback[])[]> = {
   "/projects/*": [
-    // TODO: we currently don't prefetch metrics as they are pretty slow to fetch
-    // ([_, projectId]) => (useAdminApp(projectId) as any)[stackAppInternalsSymbol].useMetrics(false),
+    ([_, projectId]) => {
+      (useAdminApp(projectId) as any)[stackAppInternalsSymbol].useMetrics(false);
+    },
+    ([_, projectId]) => {
+      useAdminApp(projectId).useUsers({ limit: 1 });
+    },
   ],
   "/projects/*/**": [
     ([_, projectId]) => {
@@ -20,8 +24,12 @@ const urlPrefetchers: Record<string, ((match: RegExpMatchArray, query: URLSearch
     },
   ],
   "/projects/*/users": [
-    // TODO: we currently don't prefetch metrics as they are pretty slow to fetch
-    // ([_, projectId]) => (useAdminApp(projectId) as any)[stackAppInternalsSymbol].useMetrics(),
+    ([_, projectId]) => {
+      (useAdminApp(projectId) as any)[stackAppInternalsSymbol].useMetrics(false);
+    },
+    ([_, projectId]) => {
+      (useAdminApp(projectId) as any)[stackAppInternalsSymbol].useMetrics(true);
+    },
     ([_, projectId]) => {
       useAdminApp(projectId).useUsers({ limit: 1 });
     },
